requirements.txt
Django
Pillow
psycopg2
djangorestframework
django-filter